@@ -1,14 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
+    
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
+        <RollForward>Major</RollForward>
     </PropertyGroup>
-
+    
     <PropertyGroup>
         <NuspecFile>Contentless.nuspec</NuspecFile>
     </PropertyGroup>
-
+    
     <ItemGroup>
         <PackageReference Include="MonoGame.Framework.Content.Pipeline" Version="3.8.0.1641">
             <PrivateAssets>All</PrivateAssets>
