@@ -1,20 +1,20 @@
 using System;
 using Microsoft.Xna.Framework.Content.Pipeline;
 
-namespace Contentless {
-    public class ImporterInfo {
+namespace Contentless; 
 
-        public readonly ContentImporterAttribute Importer;
-        public readonly Type Type;
+public class ImporterInfo {
 
-        public ImporterInfo(ContentImporterAttribute importer, Type type) {
-            this.Importer = importer;
-            this.Type = type;
-        }
+    public readonly ContentImporterAttribute Importer;
+    public readonly Type Type;
 
-        public override string ToString() {
-            return this.Type.Name;
-        }
-
+    public ImporterInfo(ContentImporterAttribute importer, Type type) {
+        this.Importer = importer;
+        this.Type = type;
     }
+
+    public override string ToString() {
+        return this.Type.Name;
+    }
+
 }
